move mae

+++ lib/data/dataloading.py
@@ -107,7 +107,7 @@ def load_raw(raw_dir, winsize, n_hours=None, sessions=None, chunk_len_hrs=5, tes
         accs.append(pad_for_windowing(torch.Tensor(acc[['x_acc','y_acc','z_acc']].values), winsize))
 
     # Concatenate all sessions and split into chunks of length chunk_len_hrs
-    chunk_len = chunk_len_hrs * 60 * 60 * 100 # number of samples at 100 Hz to get chunk_len_hrs hours
+    chunk_len = int(chunk_len_hrs * 60 * 60 * 100) # number of samples at 100 Hz to get chunk_len_hrs hours
     all_acc = torch.cat(accs, axis=0)
     all_acc = all_acc[:len(all_acc) - len(all_acc) % chunk_len] # cut off very last part
     all_acc = all_acc.view(-1, chunk_len, 3)
